requirements.txt
aiohttp
Pillow
crawl4ai
